% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeGeneSet.R
\name{getJaccardMatrix,sparseMatrix-method}
\alias{getJaccardMatrix,sparseMatrix-method}
\title{Get pairwise Jaccard similarity between genesets}
\usage{
\S4method{getJaccardMatrix}{sparseMatrix}(x)

\S4method{getJaccardMatrix}{sparseMatrix}(x)
}
\arguments{
\item{x}{collection of gene sets}
}
\description{
Get pairwise Jaccard similarity between genesets

Get pairwise Jaccard similarity between genesets
}
