% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeGeneSet.R
\name{getJaccardMatrix}
\alias{getJaccardMatrix}
\title{Get pairwise Jaccard similarity between genesets}
\usage{
getJaccardMatrix(x)

getJaccardMatrix(x)
}
\arguments{
\item{x}{collection of gene sets}
}
\description{
Get pairwise Jaccard similarity between genesets

Get pairwise Jaccard similarity between genesets
}
