% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeGeneSet.R
\name{heatplot}
\alias{heatplot}
\title{Plot gene sets in two dimensions}
\usage{
heatplot(gsc, df_GeneMap, coef, setOverlap = 2, bicluster = TRUE, zlim = NULL)
}
\arguments{
\item{gsc}{GeneSetCollection object}

\item{df_GeneMap}{result from getGeneMapping()}

\item{coef}{indicate contrast column in df_GeneMap}

\item{setOverlap}{keep genes present in at least this many genes sets}

\item{bicluster}{apply clustering to rows and columns before plotting}

\item{zlim}{scalar. specify max value for color scale}
}
\description{
Plot gene sets in two dimensions based on Jaccard distance
}
