% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinnacle.R
\name{pinnacle}
\alias{pinnacle}
\title{Evaluate Gene Set Analysis with Pinnacle}
\usage{
pinnacle(
  fit,
  coef,
  geneSets,
  data,
  formula = ~1,
  setSize = c(10, 5000),
  quiet = FALSE,
  ...
)
}
\arguments{
\item{fit}{regression model fit by \code{limma::lmFit()} or \code{variancePartition::dream()}}

\item{coef}{indicate coefficient or contrast to be extracted from \code{fit} using \code{topTable}}

\item{geneSets}{\code{GeneSetCollection} from \link{GSEABase}}

\item{data}{\code{data.frame} storing properties of each gene, with rownames being gene names}

\item{formula}{formula specifying covariates in regression using t-statistics as response}

\item{setSize}{array of two elements specifying the min and max number of genes allowed in a gene set. Only gene sets satisfying these criteria are retained}

\item{quiet}{suppresss messages}

\item{...}{other arguments passed to \code{lm_each_eclairs()} and then \link{lm}}
}
\value{
\code{data.frame} with results for each gene set
}
\description{
Perform Gene Set Analysis (GSA) by comparing t-statistics from a given gene set to genome-wide t-statistics, while accounting for co-expression structure
}
\details{

}
\examples{

}
