% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeGeneSet.R
\name{plot_mds}
\alias{plot_mds}
\title{Plot gene sets in two dimensions}
\usage{
plot_mds(gsc, res, value = "-log10(p.shift)", zlim = c(0, NA))
}
\arguments{
\item{gsc}{GeneSetCollection object}

\item{res}{data.frame with score for each gene set}

\item{value}{indicate which columns of res to use to color points}

\item{zlim}{limits of colors.  c(0,NA) indicates a range from 0 to the largest observed value}
}
\description{
Plot gene sets in two dimensions based on Jaccard distance
}
