% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeGeneSet.R
\name{recodeToDataTable}
\alias{recodeToDataTable}
\alias{recodeToDataTable,GeneSetCollection-method}
\title{Recode GeneSetCollection to data.table}
\usage{
recodeToDataTable(gsc)

\S4method{recodeToDataTable}{GeneSetCollection}(gsc)
}
\arguments{
\item{gsc}{GeneSetCollection object}
}
\description{
Recode GeneSetCollection to data.table

Recode GeneSetCollection to data.table
}
