% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeGeneSet.R
\name{recodeToSparseMatrix,GeneSetCollection-method}
\alias{recodeToSparseMatrix,GeneSetCollection-method}
\title{Recode GeneSetCollection to sparseMatrix}
\usage{
\S4method{recodeToSparseMatrix}{GeneSetCollection}(gsc, geneNames)
}
\arguments{
\item{gsc}{GeneSetCollection object}

\item{geneNames}{order of gene names in result}
}
\description{
Recode GeneSetCollection to sparseMatrix
}
